+++ src/lib/pages/module-details/index.tsx
@@ -25,6 +25,7 @@ import {
   useModuleId,
   useAccountModules,
   useModuleHistoriesCount,
+  useModuleDetailsQuery,
 } from "lib/services/moduleService";
 import { useModuleTxsCount } from "lib/services/txService";
 import type { MoveAccountAddr } from "lib/types";
@@ -96,6 +97,9 @@ export const ModuleDetailsBody = ({ moduleData }: ModuleDetailsBodyProps) => {
     moduleData.moduleName,
     moduleData.address
   );
+
+  const { data: moduleDetails, isLoading: moduleDetailsLoading } =
+    useModuleDetailsQuery(moduleId);
 
   const { data: verificationData } = useVerifyModule({
     address: moduleData.address,
@@ -251,7 +255,12 @@ export const ModuleDetailsBody = ({ moduleData }: ModuleDetailsBodyProps) => {
                   </Flex>
                 ))}
               </Flex>
-              <ModuleInfo {...moduleData} verificationData={verificationData} />
+              <ModuleInfo
+                upgradePolicy={moduleData.upgradePolicy}
+                moduleDetails={moduleDetails}
+                verificationData={verificationData}
+                isLoading={moduleDetailsLoading}
+              />
               <Flex flexDirection="column" mt={6}>
                 <Heading
                   as="h6"