+++ src/lib/services/moduleService.ts
@@ -13,6 +13,7 @@ import {
   useMoveConfig,
 } from "lib/app-provider";
 import {
+  getModuleDetailsQueryDocument,
   getModuleHistoriesCountQueryDocument,
   getModuleHistoriesQueryDocument,
   getModuleIdByNameAndVmAddressQueryDocument,
@@ -34,7 +35,9 @@ import type { ModuleHistory } from "lib/types/module";
 import {
   bech32AddressToHex,
   parseDate,
+  parseDateOpt,
   parseJsonABI,
+  parseTxHashOpt,
   splitViewExecuteFunctions,
   truncate,
   unpadHexAddress,
@@ -321,6 +324,50 @@ export const useModuleHistoriesCount = (moduleId: Option<Nullable<number>>) => {
     {
       enabled: Boolean(moduleId),
       retry: 1,
+      refetchOnWindowFocus: false,
+    }
+  );
+};
+
+export interface ModuleDetailsQueryResponse {
+  publisherVmAddress: HexAddr;
+  createdHeight: Option<number>;
+  createdTime: Option<Date>;
+  initTxHash: Option<string>;
+  initProposalId: Option<number>;
+  initProposalTitle: Option<string>;
+}
+
+export const useModuleDetailsQuery = (
+  moduleId: Option<Nullable<number>>
+): UseQueryResult<ModuleDetailsQueryResponse> => {
+  const { indexerGraphClient } = useCelatoneApp();
+
+  const queryFn = async () => {
+    if (!moduleId) throw new Error("Module id not found");
+    return indexerGraphClient
+      .request(getModuleDetailsQueryDocument, { moduleId })
+      .then<ModuleDetailsQueryResponse>(({ modules }) => {
+        const target = modules[0];
+        return {
+          publisherVmAddress: target?.publisher_vm_address
+            .vm_address as HexAddr,
+          createdHeight: target?.module_histories?.at(0)?.block.height,
+          createdTime: parseDateOpt(
+            target?.module_histories?.at(0)?.block.timestamp
+          ),
+          initTxHash: parseTxHashOpt(target?.publish_transaction?.hash),
+          initProposalId: target?.module_proposals.at(0)?.proposal.id,
+          initProposalTitle: target?.module_proposals.at(0)?.proposal.title,
+        };
+      });
+  };
+
+  return useQuery(
+    [CELATONE_QUERY_KEYS.MODULE_DETAILS, indexerGraphClient, moduleId],
+    queryFn,
+    {
+      enabled: Boolean(moduleId),
       refetchOnWindowFocus: false,
     }
   );